ANDROID-11209 Move PasswordVisibleIcon() into the library module
